Use the settings declared in the ConfigManager.

+++ Source/Storm-Logger/include/LoggerManager.cpp
@@ -43,17 +43,54 @@ Storm::LoggerManager::~LoggerManager()
 
 void Storm::LoggerManager::initialize_Implementation()
 {
+	std::vector<std::filesystem::path> logsToBeRemoved;
+
 	std::unique_lock<std::mutex> lock{ _loggerMutex };
 	_isRunning = true;
 
 	const Storm::IConfigManager* configMgr = Storm::SingletonHolder::instance().getFacet<Storm::IConfigManager>();
 	assert(configMgr != nullptr && "Config Manager should be alive when we initialize the logger!");
 
+	_level = configMgr->getLogLevel();
+	int removeLogOlderThanDay = configMgr->getRemoveLogOlderThanDaysCount();
+
 	const std::string &logFileName = configMgr->getLogFileName();
 	if (!logFileName.empty())
 	{
-		const std::filesystem::path logFilePath = std::filesystem::path{ configMgr->getTemporaryPath() } / logFileName;
-		std::filesystem::remove(logFilePath);
+		const std::filesystem::path logExtension{ ".log" };
+
+		const std::string &logFolderPathStr = configMgr->getLogFolderPath();
+		const std::filesystem::path logFolderPath = std::filesystem::path{ logFolderPathStr.empty() ? configMgr->getTemporaryPath() : logFolderPathStr };
+		std::filesystem::create_directories(logFolderPath);
+
+		std::filesystem::path logFilePath = logFolderPath / logFileName;
+		logFilePath.replace_extension(logExtension);
+
+		if (configMgr->getShouldOverrideOldLog())
+		{
+			std::filesystem::remove(logFilePath);
+		}
+		else
+		{
+			std::ofstream{ _logFilePath, std::ios_base::out | std::ios_base::app } << "\n\n\n---------------------------------------------\n\n\n";
+		}
+
+		if (removeLogOlderThanDay > 0)
+		{
+			const auto threadhold = std::filesystem::file_time_type::clock::now() - std::chrono::hours{ 24 * removeLogOlderThanDay };
+
+			for (const std::filesystem::directory_entry &logsFile : std::filesystem::recursive_directory_iterator{ logFolderPath })
+			{
+				if (logsFile.path().extension() == logExtension && (logsFile.is_character_file() || logsFile.is_regular_file()))
+				{
+					auto writeTime = logsFile.last_write_time();
+					if (writeTime <= threadhold)
+					{
+						logsToBeRemoved.emplace_back(logsFile.path());
+					}
+				}
+			}
+		}
 
 		_logFilePath = logFilePath.string();
 	}
@@ -94,6 +131,24 @@ void Storm::LoggerManager::initialize_Implementation()
 
 	// Time to log everything we wanted inside this init method.
 	LOG_ALWAYS << "Set the log level to '" << Storm::LogItem::parseLogLevel(_level) << "'. It means we wont log message under this level.";
+	
+	const std::size_t toBeRemovedCount = logsToBeRemoved.size();
+	if (toBeRemovedCount > 0)
+	{
+		LOG_COMMENT << "We would remove " << toBeRemovedCount << " log file (were detected older than " << removeLogOlderThanDay << " days from now).";
+		for (const auto &logFilePathToRemove : logsToBeRemoved)
+		{
+			if (std::filesystem::remove(logFilePathToRemove))
+			{
+				LOG_DEBUG << logFilePathToRemove << " was removed successfully";
+			}
+			else
+			{
+				LOG_DEBUG << "Could not remove " << logFilePathToRemove;
+			}
+		}
+	}
+
 	// Wait for the thread to start before continuing.
 	lock.lock();
 	syncTmp.wait(lock, [&canLeaveTmp]()
@@ -137,13 +192,16 @@ void Storm::LoggerManager::writeLogs(const LogArray &logArray) const
 		std::ofstream logFile{ _logFilePath, std::ios_base::out | std::ios_base::app };
 		for (const auto &logItem : logArray)
 		{
-			const std::string fullLogMsg{ logItem };
-			std::cout << fullLogMsg;
-			logFile << fullLogMsg;
-
-			if (debuggerAttached)
+			if (logItem._level >= _level)
 			{
-				logToVisualStudioOutput(fullLogMsg);
+				const std::string fullLogMsg{ logItem };
+				std::cout << fullLogMsg;
+				logFile << fullLogMsg;
+
+				if (debuggerAttached)
+				{
+					logToVisualStudioOutput(fullLogMsg);
+				}
 			}
 		}
 	}
@@ -151,11 +209,14 @@ void Storm::LoggerManager::writeLogs(const LogArray &logArray) const
 	{
 		for (const auto &logItem : logArray)
 		{
-			const std::string fullLogMsg{ logItem };
-			std::cout << fullLogMsg;
-			if (debuggerAttached)
+			if (logItem._level >= _level)
 			{
-				logToVisualStudioOutput(fullLogMsg);
+				const std::string fullLogMsg{ logItem };
+				std::cout << fullLogMsg;
+				if (debuggerAttached)
+				{
+					logToVisualStudioOutput(fullLogMsg);
+				}
 			}
 		}
 	}
